joypixels: Switch to general JoyPixels Free License
Our references to the Free License have never been updated. I've
received permission to link to the general license instead. This means
the license version is no longer pinned, though that does not seem worse
than being out of sync with the actual font and they don't seem to have
versioned references.

If the license is updated again before we receive new links for the font
files users will be prompted to agree to a newer version of the license
than the font. On the JoyPixels website if you navigate to older font
versions, which are still available for download, you're also pointed to
the newest version of the license. So we aren't doing worse.

Note, although it's the JoyPixels Free License, this is actually an
unfree license because it prohibits commercial use.